+++ lib/valideizer/core.rb
@@ -1,29 +1,43 @@
 require 'valideizer/rules'
 require 'valideizer/error_printer'
 require 'valideizer/caster'
+require 'valideizer/rules_checker'
 
 module Valideizer
   class Core
     include Valideizer::Rules
+    include Valideizer::RulesChecker
     include Valideizer::Caster
     include Valideizer::ErrorPrinter
 
-    def initialize
+    attr_accessor :rules
+
+    def initialize(autocast: true)
+      @autocast = autocast
       @rules = {}
       reinit_errrors
     end
 
+    # Cleanes rules and errors
+    def clean!
+      @rules = {}
+      reinit_errrors
+    end
+
+    # Adds new rule for validation
     def add_rule(param, *rules)
+      check_rule_set(rules[0])
       @rules[param.to_s] = rules[0]
     end
 
     alias valideize add_rule
 
+    # Validates and recasts params
     def valideized?(params)
       pre_process params
 
       params.each do |param, value|
-        next unless nil_check(param, value)
+        next unless null_check(param, value)
         @rules[param.to_s].each do |type, constraint|
           begin
             push_error(param, value, type, constraint) unless validate(value, type, constraint)
@@ -41,6 +55,7 @@ module Valideizer
       end
     end
 
+    # Prints error messages
     def errors
       build_error_messages
       @error_messages
@@ -80,17 +95,17 @@ module Valideizer
       end
     end
 
-    def nil_check(param, value)
-      if !value.nil? || value.nil? && has_allow_nil_rule(param)
+    def null_check(param, value)
+      if !value.nil? || value.nil? && has_allow_null_rule(param)
         true
       else
-        push_error(param, :nil, nil, nil)
+        push_error(param, :null, nil, nil)
         false
       end
     end
 
-    def has_allow_nil_rule(param)
-      nil_rule = @rules[param.to_s]&.find { |r, _c| r == :nil }
+    def has_allow_null_rule(param)
+      nil_rule = @rules[param.to_s]&.find { |r, _c| r == :null }
       nil_rule.nil? ? false : nil_rule.last
     end
 
@@ -102,36 +117,41 @@ module Valideizer
         type_rule = rules.find { |r, _c| r == :type }
         recast(params, key, type_rule.last) if type_rule
 
+        datetime_rule = rules.find { |r, _c| r == :format }
+        params[key] = cast_to_time_with_format(params[key],
+          datetime_rule.last) if datetime_rule
+
         regexp_rule = rules.find { |r, _c| r == :regexp }
         regexp_groups_substitution(params, key, regexp_rule.last) if regexp_rule
       end
     end
 
-
     def recast(params, key, type)
       value = params[key]
       params[key] = case type
-                    when :json    then cast_from_json value
-                    when :bool    then cast_to_bool value
-                    when :float   then cast_to_float value
-                    when :integer then cast_to_integer value
-                    else value
-                    end
+        when :json     then cast_from_json  value
+        when :boolean  then cast_to_boolean value
+        when :float    then cast_to_float   value
+        when :integer  then cast_to_integer value
+        when :datetime then cast_to_time    value
+        else value
+      end
     end
 
     def regexp_groups_substitution(params, key, regexp)
       value = params[key]
-
       matched = value.match regexp
       return if matched.nil?
 
       params[key] = if matched.named_captures.any?
-                      matched.named_captures
-                    elsif matched.captures.count > 1
-                      matched.captures
-                    elsif matched.captures.count == 1
-                      matched.captures[0]
-                    end
+        matched.named_captures
+      elsif matched.captures.count > 1
+        matched.captures
+      elsif matched.captures.count == 1
+        matched.captures[0]
+      else
+        value
+      end
     end
 
     def build_error_messages